LPD-52167 Move hook to new validation file
